requirements.txt
# Phase 0: CLI Voice Loop
sarvamai>=0.1.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
webrtcvad>=2.0.10
soundfile>=0.12.1
numpy>=1.24.0
